Move txn hash functions from the account pkg to the hash pkg
