+++ hash/hash.go
@@ -4,10 +4,23 @@ import (
 	"errors"
 	"slices"
 
+	"github.com/NethermindEth/juno/core/crypto"
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/starknet.go/contracts"
 	"github.com/NethermindEth/starknet.go/curve"
 	internalUtils "github.com/NethermindEth/starknet.go/internal/utils"
+	"github.com/NethermindEth/starknet.go/rpc"
+)
+
+var (
+	PREFIX_TRANSACTION    = new(felt.Felt).SetBytes([]byte("invoke"))
+	PREFIX_DECLARE        = new(felt.Felt).SetBytes([]byte("declare"))
+	PREFIX_DEPLOY_ACCOUNT = new(felt.Felt).SetBytes([]byte("deploy_account"))
+)
+
+var (
+	ErrNotAllParametersSet = errors.New("not all neccessary parameters have been set")
+	ErrFeltToBigInt        = errors.New("felt to BigInt error")
 )
 
 // CalculateDeprecatedTransactionHashCommon calculates the transaction hash common to be used in the StarkNet network - a unique identifier of the transaction.
@@ -284,4 +297,391 @@ func hashCasmClassEntryPointByType(entryPoint []contracts.CasmEntryPoint) *felt.
 		flattened = append(flattened, elt.Selector, new(felt.Felt).SetUint64(uint64(elt.Offset)), builtInHash)
 	}
 	return curve.PoseidonArray(flattened...)
+}
+
+// TransactionHashInvokeV0 calculates the transaction hash for a invoke V0 transaction.
+//
+// Parameters:
+//   - txn: The invoke V0 transaction to calculate the hash for
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashInvokeV0(txn *rpc.InvokeTxnV0, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v0_deprecated_hash_calculation
+	if txn.Version == "" || len(txn.Calldata) == 0 || txn.MaxFee == nil || txn.EntryPointSelector == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	calldataHash := curve.PedersenArray(txn.Calldata...)
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+	return CalculateDeprecatedTransactionHashCommon(
+		PREFIX_TRANSACTION,
+		txnVersionFelt,
+		txn.ContractAddress,
+		txn.EntryPointSelector,
+		calldataHash,
+		txn.MaxFee,
+		chainId,
+		[]*felt.Felt{},
+	), nil
+}
+
+// TransactionHashInvokeV1 calculates the transaction hash for a invoke V1 transaction.
+//
+// Parameters:
+//   - txn: The invoke V1 transaction to calculate the hash for
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashInvokeV1(txn *rpc.InvokeTxnV1, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v1_deprecated_hash_calculation
+	if txn.Version == "" || len(txn.Calldata) == 0 || txn.Nonce == nil || txn.MaxFee == nil || txn.SenderAddress == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	calldataHash := curve.PedersenArray(txn.Calldata...)
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+	return CalculateDeprecatedTransactionHashCommon(
+		PREFIX_TRANSACTION,
+		txnVersionFelt,
+		txn.SenderAddress,
+		&felt.Zero,
+		calldataHash,
+		txn.MaxFee,
+		chainId,
+		[]*felt.Felt{txn.Nonce},
+	), nil
+}
+
+// TransactionHashInvokeV3 calculates the transaction hash for a invoke V3 transaction.
+//
+// Parameters:
+//   - txn: The invoke V3 transaction to calculate the hash for
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashInvokeV3(txn *rpc.InvokeTxnV3, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://github.com/starknet-io/SNIPs/blob/main/SNIPS/snip-8.md#protocol-changes
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v3_hash_calculation
+	if txn.Version == "" || txn.ResourceBounds == (rpc.ResourceBoundsMapping{}) || len(txn.Calldata) == 0 || txn.Nonce == nil || txn.SenderAddress == nil || txn.PayMasterData == nil || txn.AccountDeploymentData == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+	DAUint64, err := dataAvailabilityMode(txn.FeeMode, txn.NonceDataMode)
+	if err != nil {
+		return nil, err
+	}
+	tipUint64, err := txn.Tip.ToUint64()
+	if err != nil {
+		return nil, err
+	}
+	tipAndResourceHash, err := tipAndResourcesHash(tipUint64, txn.ResourceBounds)
+	if err != nil {
+		return nil, err
+	}
+	return crypto.PoseidonArray(
+		PREFIX_TRANSACTION,
+		txnVersionFelt,
+		txn.SenderAddress,
+		tipAndResourceHash,
+		crypto.PoseidonArray(txn.PayMasterData...),
+		chainId,
+		txn.Nonce,
+		new(felt.Felt).SetUint64(DAUint64),
+		crypto.PoseidonArray(txn.AccountDeploymentData...),
+		crypto.PoseidonArray(txn.Calldata...),
+	), nil
+}
+
+// TransactionHashDeclareV1 calculates the transaction hash for a declare V1 transaction.
+//
+// Parameters:
+//   - txn: The declare V1 transaction to calculate the hash for
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashDeclareV1(txn *rpc.DeclareTxnV1, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v1_deprecated_hash_calculation_2
+	if txn.SenderAddress == nil || txn.Version == "" || txn.ClassHash == nil || txn.MaxFee == nil || txn.Nonce == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	calldataHash := curve.PedersenArray(txn.ClassHash)
+
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+	return CalculateDeprecatedTransactionHashCommon(
+		PREFIX_DECLARE,
+		txnVersionFelt,
+		txn.SenderAddress,
+		&felt.Zero,
+		calldataHash,
+		txn.MaxFee,
+		chainId,
+		[]*felt.Felt{txn.Nonce},
+	), nil
+}
+
+// TransactionHashDeclareV2 calculates the transaction hash for a declare V2 transaction.
+//
+// Parameters:
+//   - txn: The declare V2 transaction to calculate the hash for
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashDeclareV2(txn *rpc.DeclareTxnV2, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v2_deprecated_hash_calculation
+	if txn.CompiledClassHash == nil || txn.SenderAddress == nil || txn.Version == "" || txn.ClassHash == nil || txn.MaxFee == nil || txn.Nonce == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	calldataHash := curve.PedersenArray(txn.ClassHash)
+
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+
+	return CalculateDeprecatedTransactionHashCommon(
+		PREFIX_DECLARE,
+		txnVersionFelt,
+		txn.SenderAddress,
+		&felt.Zero,
+		calldataHash,
+		txn.MaxFee,
+		chainId,
+		[]*felt.Felt{txn.Nonce, txn.CompiledClassHash},
+	), nil
+}
+
+// TransactionHashDeclareV3 calculates the transaction hash for a declare V3 transaction.
+//
+// Parameters:
+//   - txn: The declare V3 transaction to calculate the hash for
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashDeclareV3(txn *rpc.DeclareTxnV3, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v3_hash_calculation_2
+	// https://github.com/starknet-io/SNIPs/blob/main/SNIPS/snip-8.md#protocol-changes
+	if txn.Version == "" || txn.ResourceBounds == (rpc.ResourceBoundsMapping{}) || txn.Nonce == nil || txn.SenderAddress == nil || txn.PayMasterData == nil || txn.AccountDeploymentData == nil ||
+		txn.ClassHash == nil || txn.CompiledClassHash == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+	DAUint64, err := dataAvailabilityMode(txn.FeeMode, txn.NonceDataMode)
+	if err != nil {
+		return nil, err
+	}
+	tipUint64, err := txn.Tip.ToUint64()
+	if err != nil {
+		return nil, err
+	}
+
+	tipAndResourceHash, err := tipAndResourcesHash(tipUint64, txn.ResourceBounds)
+	if err != nil {
+		return nil, err
+	}
+	return crypto.PoseidonArray(
+		PREFIX_DECLARE,
+		txnVersionFelt,
+		txn.SenderAddress,
+		tipAndResourceHash,
+		crypto.PoseidonArray(txn.PayMasterData...),
+		chainId,
+		txn.Nonce,
+		new(felt.Felt).SetUint64(DAUint64),
+		crypto.PoseidonArray(txn.AccountDeploymentData...),
+		txn.ClassHash,
+		txn.CompiledClassHash,
+	), nil
+}
+
+// TransactionHashBroadcastDeclareV3 calculates the transaction hash for a broadcast declare V3 transaction.
+//
+// Parameters:
+//   - txn: The broadcast declare V3 transaction to calculate the hash for
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashBroadcastDeclareV3(txn *rpc.BroadcastDeclareTxnV3, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v3_hash_calculation_2
+	// https://github.com/starknet-io/SNIPs/blob/main/SNIPS/snip-8.md#protocol-changes
+	if txn.Version == "" || txn.ResourceBounds == (rpc.ResourceBoundsMapping{}) || txn.Nonce == nil || txn.SenderAddress == nil || txn.PayMasterData == nil || txn.AccountDeploymentData == nil ||
+		txn.ContractClass == nil || txn.CompiledClassHash == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+	DAUint64, err := dataAvailabilityMode(txn.FeeMode, txn.NonceDataMode)
+	if err != nil {
+		return nil, err
+	}
+	tipUint64, err := txn.Tip.ToUint64()
+	if err != nil {
+		return nil, err
+	}
+
+	tipAndResourceHash, err := tipAndResourcesHash(tipUint64, txn.ResourceBounds)
+	if err != nil {
+		return nil, err
+	}
+	return crypto.PoseidonArray(
+		PREFIX_DECLARE,
+		txnVersionFelt,
+		txn.SenderAddress,
+		tipAndResourceHash,
+		crypto.PoseidonArray(txn.PayMasterData...),
+		chainId,
+		txn.Nonce,
+		new(felt.Felt).SetUint64(DAUint64),
+		crypto.PoseidonArray(txn.AccountDeploymentData...),
+		ClassHash(txn.ContractClass),
+		txn.CompiledClassHash,
+	), nil
+}
+
+// TransactionHashDeployAccountV1 calculates the transaction hash for a deploy account V1 transaction.
+//
+// Parameters:
+//   - txn: The deploy account V1 transaction to calculate the hash for
+//   - contractAddress: The contract address as parameters as a *felt.Felt
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashDeployAccountV1(txn *rpc.DeployAccountTxn, contractAddress, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v1_deprecated_hash_calculation_3
+	calldata := []*felt.Felt{txn.ClassHash, txn.ContractAddressSalt}
+	calldata = append(calldata, txn.ConstructorCalldata...)
+	calldataHash := curve.PedersenArray(calldata...)
+
+	versionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+
+	return CalculateDeprecatedTransactionHashCommon(
+		PREFIX_DEPLOY_ACCOUNT,
+		versionFelt,
+		contractAddress,
+		&felt.Zero,
+		calldataHash,
+		txn.MaxFee,
+		chainId,
+		[]*felt.Felt{txn.Nonce},
+	), nil
+}
+
+// TransactionHashDeployAccountV3 calculates the transaction hash for a deploy account V3 transaction.
+//
+// Parameters:
+//   - txn: The deploy account V3 transaction to calculate the hash for
+//   - contractAddress: The contract address as parameters as a *felt.Felt
+//   - chainId: The chain ID as a *felt.Felt
+//
+// Returns:
+//   - *felt.Felt: the calculated transaction hash
+//   - error: an error if any
+func TransactionHashDeployAccountV3(txn *rpc.DeployAccountTxnV3, contractAddress, chainId *felt.Felt) (*felt.Felt, error) {
+	// https://docs.starknet.io/architecture-and-concepts/network-architecture/transactions/#v3_hash_calculation_3
+	if txn.Version == "" || txn.ResourceBounds == (rpc.ResourceBoundsMapping{}) || txn.Nonce == nil || txn.PayMasterData == nil {
+		return nil, ErrNotAllParametersSet
+	}
+
+	txnVersionFelt, err := new(felt.Felt).SetString(string(txn.Version))
+	if err != nil {
+		return nil, err
+	}
+	DAUint64, err := dataAvailabilityMode(txn.FeeMode, txn.NonceDataMode)
+	if err != nil {
+		return nil, err
+	}
+	tipUint64, err := txn.Tip.ToUint64()
+	if err != nil {
+		return nil, err
+	}
+	tipAndResourceHash, err := tipAndResourcesHash(tipUint64, txn.ResourceBounds)
+	if err != nil {
+		return nil, err
+	}
+	return crypto.PoseidonArray(
+		PREFIX_DEPLOY_ACCOUNT,
+		txnVersionFelt,
+		contractAddress,
+		tipAndResourceHash,
+		crypto.PoseidonArray(txn.PayMasterData...),
+		chainId,
+		txn.Nonce,
+		new(felt.Felt).SetUint64(DAUint64),
+		crypto.PoseidonArray(txn.ConstructorCalldata...),
+		txn.ClassHash,
+		txn.ContractAddressSalt,
+	), nil
+}
+
+func tipAndResourcesHash(tip uint64, resourceBounds rpc.ResourceBoundsMapping) (*felt.Felt, error) {
+	l1Bytes, err := resourceBounds.L1Gas.Bytes(rpc.ResourceL1Gas)
+	if err != nil {
+		return nil, err
+	}
+	l2Bytes, err := resourceBounds.L2Gas.Bytes(rpc.ResourceL2Gas)
+	if err != nil {
+		return nil, err
+	}
+	l1DataGasBytes, err := resourceBounds.L1DataGas.Bytes(rpc.ResourceL1DataGas)
+	if err != nil {
+		return nil, err
+	}
+	l1Bounds := new(felt.Felt).SetBytes(l1Bytes)
+	l2Bounds := new(felt.Felt).SetBytes(l2Bytes)
+	l1DataGasBounds := new(felt.Felt).SetBytes(l1DataGasBytes)
+	return crypto.PoseidonArray(new(felt.Felt).SetUint64(tip), l1Bounds, l2Bounds, l1DataGasBounds), nil
+}
+
+func dataAvailabilityMode(feeDAMode, nonceDAMode rpc.DataAvailabilityMode) (uint64, error) {
+	const dataAvailabilityModeBits = 32
+	fee64, err := feeDAMode.UInt64()
+	if err != nil {
+		return 0, err
+	}
+	nonce64, err := nonceDAMode.UInt64()
+	if err != nil {
+		return 0, err
+	}
+	return fee64 + nonce64<<dataAvailabilityModeBits, nil
 }
